renamed nlp_functions to nlp, pattern_match to pattern_match_main
