+++ nlp/pattern_match_main.py
@@ -28,7 +28,7 @@ TODO: incorporate negation detection
 import argparse
 
 import file_utils.array_io_utils as utils
-from nlp_functions.pattern_match_utils import (load_tags,
+from nlp.pattern_match_utils import (load_tags,
                                                pattern_match_keyword,
                                                pattern_match_window)
